Change BatchedGrandProductProof to use Gruen's optimization
The primary change in this commit is to change the BatchedGrandProductProof
trait's compute_cubic method to take a GruenSplitEqPolynomal, instead of a
SplitEqPolynomial. This allows us to compute the cubic polynomials that the eq
polynomials are used to compute by first computing quadratic polynomials, per
Gruen's optimization. Rather than computing the evaluations at points 0<=j<=3,
we instead compute the evaluations at 0 and infinity.